Hermes: Unlocking Security Analysis of Cellular Network Protocols
In this paper, we present Hermes, an end-to-end framework
to automatically generate formal representations from natural
language cellular specifications. We first develop a neural
constituency parser, NEUTREX, to process transition-relevant
texts and extract transition components (i.e., states, conditions,
and actions). We also design a domain-specific language to
translate these transition components to logical formulas by
leveraging dependency parse trees. Finally, we compile these
logical formulas to generate transitions and create the formal
model as finite state machines. To demonstrate the effective-
ness of Hermes, we evaluate it on 4G NAS, 5G NAS, and 5G
RRC specifications and obtain an overall accuracy of 81-87%,
which is a substantial improvement over the state-of-the-art.
Our security analysis of the extracted models uncovers 3 new
vulnerabilities and identifies 19 previous attacks in 4G and 5G
specifications, and 7 deviations in commercial 4G basebands